Note.rename
