+++ CHANGELOG.md
@@ -80,7 +80,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - references finds all occurrences of the same footnote in the current note.
   - completion of existing footnotes after typing `[^`, plus a create item for unresolved ids that prompts for content and inserts the definition (`obsidian.footnote_new`).
   - `:Obsidian footnotes` to view footnotes of the current note via `vim.ui.select` with preview.
-- Checking invalid filenames on file create, rename and Note.create calls.
+- `Note.rename` for scripting note renames.
+- Invalid filenames are checked on `Note.create` and `Note.rename`.
 
 ### Fixed
