+++ lua/obsidian/lsp/handlers.lua
@@ -4,7 +4,6 @@ return {
   ["initialized"] = require "obsidian.lsp.handlers.initialized",
   ["workspace/didChangeWatchedFiles"] = require "obsidian.lsp.handlers.did_change_watched_files",
   ["workspace/didRenameFiles"] = require "obsidian.lsp.handlers.did_rename_files",
-  ["workspace/didCreateFiles"] = require "obsidian.lsp.handlers.did_create_files",
   ["workspace/symbol"] = require "obsidian.lsp.handlers.workspace_symbol",
   ["workspace/executeCommand"] = require "obsidian.lsp.handlers.workspace_execute_command",
   ["textDocument/rename"] = require "obsidian.lsp.handlers.rename",
