+++ lua/obsidian/lsp/handlers/initialize.lua
@@ -42,17 +42,6 @@ local initializeResult = {
     },
     workspace = {
       fileOperations = {
-        didCreate = {
-          filters = {
-            {
-              scheme = "file",
-              pattern = {
-                glob = "**/*.md",
-                matches = "file",
-              },
-            },
-          },
-        },
         didRename = {
           filters = {
             {
