+++ lua/obsidian/lsp/handlers/rename.lua
@@ -1,12 +1,9 @@
 local obsidian = require "obsidian"
 
-local Note = obsidian.Note
 local search = obsidian.search
 local log = obsidian.log
 local api = obsidian.api
 local util = obsidian.util
-
-local M = require "obsidian.lsp.handlers._rename"
 
 ---@param params lsp.RenameParams
 return function(params, handler, _)
@@ -21,25 +18,15 @@ return function(params, handler, _)
     return log.err(err and err or "failed writing all buffers before renaming, abort")
   end
 
-  local valid, reason = Note.is_valid_filename(new_name)
-  if not valid then
-    log.err(("Invalid filename %q: %s"):format(new_name, reason))
-    return handler(nil, {})
-  end
-
   local cur_link = api.cursor_link(bufnr, params.position)
 
   local function do_rename(note)
-    local old_stem = note.path and note.path.stem or nil
-    if new_name == note.id or (old_stem and new_name == old_stem) then
-      log.info "Identical name"
-      return handler(nil, {})
-    end
-    if not M.validate(new_name, workspace_dir) then
-      log.info "Note with same name exists"
-      return handler(nil, {})
-    end
-    M.rename(note, new_name, handler, { dir = workspace_dir })
+    note:rename(new_name, { apply = false }, function(err, edit)
+      if err or not edit then
+        return handler(nil, {})
+      end
+      handler(nil, edit)
+    end)
   end
 
   if cur_link then
