+++ tests/lsp/test_did_rename_files.lua
@@ -22,43 +22,43 @@ end
 T["didRenameFiles applies reference edits without file rename"] = function()
   child.lua [[
     local handler = require "obsidian.lsp.handlers.did_rename_files"
-    local rename = require "obsidian.lsp.handlers._rename"
     local note_mod = require "obsidian.note"
     local api = require "obsidian.api"
     local Path = require "obsidian.path"
-    local old_build_edit = rename.build_edit
     local old_from_file = note_mod.from_file
     local old_confirm = api.confirm
 
     note_mod.from_file = function(path)
       return {
         path = Path.new(path),
-      }
-    end
-
-    rename.build_edit = function(note, new_name, opts, callback)
-      _G.captured_note_path = tostring(note.path)
-      _G.captured_name = new_name
-      _G.captured_old_path = opts.old_path
-      _G.captured_new_path = opts.new_path
-      _G.captured_include_file_rename = opts.include_file_rename
-      callback({
-        documentChanges = {
-          {
-            textDocument = {
-              uri = vim.uri_from_fname "/tmp/ref.md",
-              version = vim.NIL,
+        rename = function(self, new_name, opts, callback)
+          _G.captured_note_path = tostring(self.path)
+          _G.captured_name = new_name
+          _G.captured_old_path = opts.old_path
+          _G.captured_new_path = opts.new_path
+          _G.captured_include_file_rename = opts.include_file_rename
+          _G.captured_apply = opts.apply
+          _G.captured_update_buffers = opts.update_buffers
+          _G.captured_check_unique = opts.check_unique
+          callback(nil, {
+            documentChanges = {
+              {
+                textDocument = {
+                  uri = vim.uri_from_fname "/tmp/ref.md",
+                  version = vim.NIL,
+                },
+                edits = {},
+              },
             },
-            edits = {},
-          },
-        },
-      }, {
-        count = 2,
-        path_lookup = { ["/tmp/ref.md"] = true },
-        buf_list = {},
-        old_path = opts.old_path,
-        new_path = opts.new_path,
-      })
+          }, {
+            count = 2,
+            path_lookup = { ["/tmp/ref.md"] = true },
+            buf_list = {},
+            old_path = opts.old_path,
+            new_path = opts.new_path,
+          })
+        end,
+      }
     end
 
     api.confirm = function(prompt)
@@ -80,7 +80,6 @@ T["didRenameFiles applies reference edits without file rename"] = function()
       end,
     })
 
-    rename.build_edit = old_build_edit
     note_mod.from_file = old_from_file
     api.confirm = old_confirm
   ]]
@@ -90,6 +89,9 @@ T["didRenameFiles applies reference edits without file rename"] = function()
   eq("/tmp/folder/old.md", child.lua_get "captured_old_path")
   eq("/tmp/folder/new.md", child.lua_get "captured_new_path")
   eq(false, child.lua_get "captured_include_file_rename")
+  eq(false, child.lua_get "captured_apply")
+  eq(false, child.lua_get "captured_update_buffers")
+  eq(false, child.lua_get "captured_check_unique")
   eq("Update 2 reference(s) across 1 file(s) for renamed note 'new'?", child.lua_get "confirm_prompt")
   eq("workspace/applyEdit", child.lua_get "request_method")
   eq("Update renamed note references", child.lua_get "request_label")
@@ -98,11 +100,9 @@ end
 T["didRenameFiles skips applyEdit when confirmation is declined"] = function()
   child.lua [[
     local handler = require "obsidian.lsp.handlers.did_rename_files"
-    local rename = require "obsidian.lsp.handlers._rename"
     local note_mod = require "obsidian.note"
     local api = require "obsidian.api"
     local Path = require "obsidian.path"
-    local old_build_edit = rename.build_edit
     local old_from_file = note_mod.from_file
     local old_confirm = api.confirm
 
@@ -111,27 +111,26 @@ T["didRenameFiles skips applyEdit when confirmation is declined"] = function()
     note_mod.from_file = function(path)
       return {
         path = Path.new(path),
-      }
-    end
-
-    rename.build_edit = function(_, _, opts, callback)
-      callback({
-        documentChanges = {
-          {
-            textDocument = {
-              uri = vim.uri_from_fname "/tmp/ref.md",
-              version = vim.NIL,
+        rename = function(_, _, opts, callback)
+          callback(nil, {
+            documentChanges = {
+              {
+                textDocument = {
+                  uri = vim.uri_from_fname "/tmp/ref.md",
+                  version = vim.NIL,
+                },
+                edits = {},
+              },
             },
-            edits = {},
-          },
-        },
-      }, {
-        count = 1,
-        path_lookup = { ["/tmp/ref.md"] = true },
-        buf_list = {},
-        old_path = opts.old_path,
-        new_path = opts.new_path,
-      })
+          }, {
+            count = 1,
+            path_lookup = { ["/tmp/ref.md"] = true },
+            buf_list = {},
+            old_path = opts.old_path,
+            new_path = opts.new_path,
+          })
+        end,
+      }
     end
 
     api.confirm = function(prompt)
@@ -152,7 +151,6 @@ T["didRenameFiles skips applyEdit when confirmation is declined"] = function()
       end,
     })
 
-    rename.build_edit = old_build_edit
     note_mod.from_file = old_from_file
     api.confirm = old_confirm
   ]]
@@ -164,11 +162,9 @@ end
 T["didRenameFiles skips confirmation when auto_update is enabled"] = function()
   child.lua [[
     local handler = require "obsidian.lsp.handlers.did_rename_files"
-    local rename = require "obsidian.lsp.handlers._rename"
     local note_mod = require "obsidian.note"
     local api = require "obsidian.api"
     local Path = require "obsidian.path"
-    local old_build_edit = rename.build_edit
     local old_from_file = note_mod.from_file
     local old_confirm = api.confirm
 
@@ -179,17 +175,16 @@ T["didRenameFiles skips confirmation when auto_update is enabled"] = function()
     note_mod.from_file = function(path)
       return {
         path = Path.new(path),
+        rename = function(_, _, _, callback)
+          callback(nil, { documentChanges = {} }, {
+            count = 0,
+            path_lookup = {},
+            buf_list = {},
+            old_path = "",
+            new_path = "",
+          })
+        end,
       }
-    end
-
-    rename.build_edit = function(_, _, _, callback)
-      callback({ documentChanges = {} }, {
-        count = 0,
-        path_lookup = {},
-        buf_list = {},
-        old_path = "",
-        new_path = "",
-      })
     end
 
     api.confirm = function()
@@ -211,7 +206,6 @@ T["didRenameFiles skips confirmation when auto_update is enabled"] = function()
     })
 
     Obsidian.opts.link.auto_update = false
-    rename.build_edit = old_build_edit
     note_mod.from_file = old_from_file
     api.confirm = old_confirm
   ]]
